Added single source of truth removed multiple variables that were holding contact data and added single source
