+++ businessContacts/Views/ContentView.swift
@@ -18,7 +18,6 @@ struct ContentView: View {
     @State private var contactToEdit: Contact?
     @State private var searchText = ""
     
-    
     var searchResults: [Contact] {
         if searchText.isEmpty {
             return contacts
