+++ businessContacts/Views/UpdateContactView.swift
@@ -1,74 +1,72 @@
 //
-//  AddNewContact.swift
+//  UpdateContactView.swift
 //  businessContacts
 //
-//  Created by Benia Morgan-Ware on 8/26/23.
+//  Created by Benia Morgan-Ware on 12/3/23.
 //
 
 import SwiftUI
 import SwiftData
 import PhotosUI
 
-struct AddNewContact: View {
+
+struct UpdateContactView: View {
     
     @Environment(\.modelContext) var context
     
     
     @Environment(\.dismiss) private var dismiss
-    @State var name: String = ""
-    @State var email: String = ""
-    @State var company: String = ""
+    @Bindable var contact: Contact
     @State private var contactAvatarItem: PhotosPickerItem?
-    @State private var contactAvatarImage: Data?
     
-    var isFormValid: Bool {
-        return !name.isEmpty && !email.isEmpty && !company.isEmpty
+    
+   var isFormValid: Bool {
+       return !contact.name!.isEmpty && !contact.email!.isEmpty && !contact.company!.isEmpty
     }
     
     var body: some View {
         NavigationStack {
             ZStack {
                 Color.gray.opacity(0.1)
-                    .ignoresSafeArea(.all)
+                    .ignoresSafeArea()
                 
                 VStack {
-                    if let contactAvatarImage {
-                       let uiImage = UIImage(data: contactAvatarImage)
+                    if contact.contactImageData != nil {
+                        let uiImage = UIImage(data: contact.contactImageData!)
                         Image(uiImage: uiImage!)
                             .resizable()
                             .scaledToFit()
                             .frame(width: 300, height: 300)
+                        
+                        Button("Remove Image", role: .destructive) {
+                            withAnimation {
+                                contactAvatarItem = nil
+                                contact.contactImageData = nil
+                            }
+                        }
+                        .buttonStyle(.bordered)
+                        .buttonBorderShape(.capsule)
                         
                     } else {
                         Image(systemName:"person.crop.circle.fill")
                             .font(.system(size: 130))
                             .foregroundColor(.gray)
                             .padding(.top, 30)
-                    }
-                    
-                    if contactAvatarImage != nil {
-                        Button("Remove Image", role: .destructive) {
-                            withAnimation {
-                                contactAvatarItem = nil
-                                contactAvatarImage = nil
-                            }
-                        }
-                        .buttonStyle(.bordered)
-                        .buttonBorderShape(.capsule)
-                    } else {
                         
                         PhotosPicker("Add Photo", selection: $contactAvatarItem, matching: .images)
                             .buttonStyle(.bordered)
                             .buttonBorderShape(.capsule)
                     }
                     
-                    Form {
-                        TextField("Full Name", text: $name)
-                        
-                        TextField("Email", text: $email)
-                            .keyboardType(.emailAddress)
-                        
-                        TextField("Company", text: $company)
+                    VStack {
+                        Form {
+                            TextField("Full Name", text: $contact.name.defaultValue(""))
+                            
+                            TextField("Email", text: $contact.email.defaultValue(""))
+                                .keyboardType(.emailAddress)
+                            
+                            TextField("Company", text: $contact.company.defaultValue(""))
+                        }
                     }
                     .padding(.top)
                     .autocorrectionDisabled()
@@ -76,37 +74,24 @@ struct AddNewContact: View {
                     Spacer()
                     
                         .toolbar {
-                            ToolbarItem(placement: .navigationBarTrailing) {
+                            ToolbarItem(placement: .confirmationAction) {
                                 Button("Done") {
                                     withAnimation {
-                                        if isFormValid {
-                                            let newContact = Contact(name: name, email: email, company: company, contactImageData: contactAvatarImage)
-                                            
-                                            context.insert(newContact)
-                                            
-                                            dismiss()
-                                        }
+                                        
+                                        dismiss()
                                     }
-                                    
                                 }
                                 .bold()
                                 .disabled(!isFormValid)
-                                
-                            }
-                            ToolbarItem(placement: .navigationBarLeading) {
-                                Button("Cancel") {
-                                    dismiss()
-                                }
-                                
                             }
                         }
                 }
-                .navigationTitle("New Contact")
+                .navigationTitle("Update Contact")
                 .navigationBarTitleDisplayMode(.inline)
                 .onChange(of: contactAvatarItem) {
                     Task {
                         if let loaded  = try? await contactAvatarItem?.loadTransferable(type: Data.self) {
-                            contactAvatarImage = loaded
+                            contact.contactImageData = loaded
                         } else {
                             print("Failed to Load Image")
                         }
@@ -118,7 +103,18 @@ struct AddNewContact: View {
     }
 }
 
+extension Binding {
+    public func defaultValue<T>(_ value: T) -> Binding<T> where Value == Optional<T> {
+        Binding<T> {
+            wrappedValue ?? value
+        } set: {
+            wrappedValue = $0
+        }
+    }
+}
+
 #Preview {
     let preview = previewContainer([Contact.self])
-    return AddNewContact().modelContainer(preview.container)
+    return UpdateContactView(contact: Contact(name: "", email: "", company: "", contactImageData: nil)).modelContainer(preview.container)
 }
+
